refactored misc. MM functionality ktmm_vmscan.c to separate files
