+++ Kbuild
@@ -2,5 +2,7 @@ EXTRA_CFLAGS += -DDEBUG
 
 obj-m += ktmm.o
 ktmm-y := src/ktmm_main.o
-ktmm-y += src/ktmm_vmscan.o
 ktmm-y += src/ktmm_hook.o
+ktmm-y += src/ktmm_vmscan.o
+ktmm-y += src/ktmm_vmhooks.o
+ktmm-y += src/ktmm_mm.o
